Feat LIC0 done
Fix issue/1
LIC0 is done and two tests done to test the correctness of LIC0

+++ src/main/java/com/example/Coordinate.java
@@ -9,11 +9,11 @@ public class Coordinate{
 		this.y = y;
 	}
 	
-	public double getX (){
+	public Integer getX (){
 		return x;
 	}
 	
-	public double getY (){
+	public Integer getY (){
 		return y;
 	}
 }
